chore: merge PersonaService into IdentityService
