+++ src/extension/background-script/IdentityService.ts
@@ -1,5 +1,6 @@
+import * as bip39 from 'bip39'
 import { queryProfilesWithQuery, personaRecordToPersona, storeAvatar, queryProfile } from '../../database'
-import { ProfileIdentifier, PersonaIdentifier } from '../../database/type'
+import { ProfileIdentifier, PersonaIdentifier, Identifier, ECKeyIdentifier } from '../../database/type'
 import type { Profile, Persona } from '../../database/Persona/types'
 import {
     queryPersonaDB,
@@ -13,12 +14,19 @@ import {
     createOrUpdateProfileDB,
     consistentPersonaDBWriteAccess,
 } from '../../database/Persona/Persona.db'
+import { queryPersona } from '../../database'
 import { OnlyRunInContext } from '@holoflows/kit/es'
+import { BackupJSONFileLatest, UpgradeBackupJSONFile } from '../../utils/type-transform/BackupFormat/JSON/latest'
+import { restoreBackup } from './WelcomeServices/restoreBackup'
+import { restoreNewIdentityWithMnemonicWord } from './WelcomeService'
+import { decodeText, decodeArrayBuffer } from '../../utils/type-transform/String-ArrayBuffer'
+import { decompressBackupFile } from '../../utils/type-transform/BackupFileShortRepresentation'
 
 OnlyRunInContext(['background', 'debugging'], 'IdentityService')
 
-// region Avatars
+//#region Avatars
 export { storeAvatar, getAvatarDataURL as queryAvatarDataURL } from '../../database'
+//#endregion
 
 //#region Profile
 export { queryProfile, queryProfilePaged } from '../../database'
@@ -74,13 +82,34 @@ export function queryMyPersonas(network?: string): Promise<Persona[]> {
             : x,
     )
 }
+export async function restoreFromObject(object: null | BackupJSONFileLatest) {
+    if (!object) return null
+    await restoreBackup(object!)
+    if (object?.personas?.length) {
+        return queryPersona(Identifier.fromString(object.personas[0].identifier, ECKeyIdentifier).unwrap())
+    }
+    return null
+}
+export async function restoreFromMnemonicWords(mnemonicWords: string, nickname: string, password: string) {
+    if (!bip39.validateMnemonic(mnemonicWords)) throw new Error('the mnemonic words are not valid')
+    const identifier = await restoreNewIdentityWithMnemonicWord(mnemonicWords, password, {
+        nickname,
+    })
+    return queryPersona(identifier)
+}
+export async function restoreFromBase64(base64: string) {
+    return restoreFromObject(JSON.parse(decodeText(decodeArrayBuffer(base64))) as BackupJSONFileLatest)
+}
+export async function restoreFromBackup(backup: string) {
+    return restoreFromObject(UpgradeBackupJSONFile(decompressBackupFile(backup)))
+}
+//#endregion
+
+//#region Profile & Persona
 /**
  * Remove an identity.
  */
 export { setupPersona, deletePersona } from '../../database'
-//#endregion
-
-//#region Profile & Persona
 export async function attachProfile(
     source: ProfileIdentifier,
     target: ProfileIdentifier | PersonaIdentifier,
